requirements.txt
flask
python-dotenv
pymongo
werkzeug
flask-cors
